Step 1

+++ unit-test/openvibe-kernel/urErrorManagerTest.cpp
@@ -114,10 +114,10 @@ TEST(error_manager_test_case, test_stress_push)
 	auto& errorManager = context->getErrorManager();
 
 	errorManager.releaseErrors();
-	const uint32_t expectedErrorCount = 10;
-	for (uint32_t i = 0; i < expectedErrorCount; ++i) { errorManager.pushError(ErrorType::Unknown, "Error"); }
+	const size_t expectedErrorCount = 10;
+	for (size_t i = 0; i < expectedErrorCount; ++i) { errorManager.pushError(ErrorType::Unknown, "Error"); }
 
-	uint32_t errorCount = 0;
+	size_t errorCount = 0;
 	auto error              = errorManager.getLastError();
 	while (error)
 	{
